Created MarcieEmbed helper class simplify - MarcieEmbed class should help with embed use and conversion methods - Helps with organization - Helps readability in `Fftcg.py`
