+++ marcie.py
@@ -1,4 +1,3 @@
-import discord
 from discord.ext import commands
 import datetime
 import logging
